requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
python-dotenv